Refactored to MVC and created the Main Window view/controller.
