+++ src/com/cbrobsto/Launcher.java
@@ -15,8 +15,8 @@ public class Launcher extends Application {
     @Override
     public void start(Stage stage) throws Exception {
 
-        Parent parent = FXMLLoader.load(getClass().getResource("view/LoginWindow.fxml"));
-        Scene scene = new Scene(parent, 510, 365);
+        Parent parent = FXMLLoader.load(getClass().getResource("view/MainWindow.fxml"));
+        Scene scene = new Scene(parent);
         stage.setScene(scene);
 
         stage.show();
