+++ src/com/cbrobsto/controller/MainWindowController.java
@@ -1,0 +1,25 @@
+package com.cbrobsto.view;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.TableView;
+import javafx.scene.control.TreeView;
+import javafx.scene.web.WebView;
+
+public class MainWindowController {
+
+    @FXML
+    private TreeView<?> emailsTreeView;
+
+    @FXML
+    private TableView<?> emailsTableView;
+
+    @FXML
+    private WebView emailWebView;
+
+    @FXML
+    void optionsAction() {
+
+    }
+
+}
+
