+++ src/com/cbrobsto/view/LoginWindow.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.layout.AnchorPane?>
 
-<AnchorPane prefHeight="365.0" prefWidth="510.0" xmlns="http://javafx.com/javafx/15.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.cbrobsto.view.LoginWindowController">
+<AnchorPane prefHeight="365.0" prefWidth="510.0" xmlns="http://javafx.com/javafx/15.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.cbrobsto.controller.LoginWindowController">
    <children>
       <Button layoutX="194.0" layoutY="252.0" mnemonicParsing="false" onAction="#loginButtonAction" prefHeight="52.0" prefWidth="119.0" text="Login" />
       <Label layoutX="23.0" layoutY="130.0" text="Email Address" />
