+++ src/com/cbrobsto/view/MainWindow.fxml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Menu?>
+<?import javafx.scene.control.MenuBar?>
+<?import javafx.scene.control.MenuItem?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.TreeView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.web.WebView?>
+
+
+<AnchorPane prefHeight="596.0" prefWidth="921.0" xmlns="http://javafx.com/javafx/15.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.cbrobsto.view.MainWindowController">
+   <children>
+      <MenuBar prefHeight="29.0" prefWidth="921.0">
+        <menus>
+          <Menu mnemonicParsing="false" text="File">
+            <items>
+              <MenuItem mnemonicParsing="false" text="Close" />
+                  <MenuItem mnemonicParsing="false" text="Open" />
+            </items>
+          </Menu>
+          <Menu mnemonicParsing="false" text="Edit">
+            <items>
+              <MenuItem mnemonicParsing="false" onAction="#optionsAction" text="Options" />
+            </items>
+          </Menu>
+          <Menu mnemonicParsing="false" text="Help">
+            <items>
+              <MenuItem mnemonicParsing="false" text="About" />
+            </items>
+          </Menu>
+        </menus>
+      </MenuBar>
+      <TreeView fx:id="emailsTreeView" layoutY="28.0" prefHeight="568.0" prefWidth="200.0" />
+      <TableView fx:id="emailsTableView" layoutX="200.0" layoutY="28.0" prefHeight="188.0" prefWidth="722.0">
+        <columns>
+          <TableColumn prefWidth="130.0" text="Sender" />
+            <TableColumn prefWidth="245.0" text="Subject" />
+          <TableColumn minWidth="0.0" prefWidth="153.0" text="Recipient" />
+            <TableColumn minWidth="0.0" prefWidth="89.0" text="Size" />
+            <TableColumn prefWidth="93.0" text="Date" />
+        </columns>
+      </TableView>
+      <WebView fx:id="emailWebView" layoutX="200.0" layoutY="216.0" prefHeight="380.0" prefWidth="722.0" />
+   </children>
+</AnchorPane>
